backend/pyproject.toml
[project]
name = "bikeroutes-backend"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
  "slowapi==0.1.9",
  "fastapi==0.115.8",
  "pydantic-settings==2.7.1",
  "python-dotenv==1.0.1",
  "uvicorn[standard]==0.34.0",
  "SQLAlchemy==2.0.36",
  "alembic==1.14.0",]

[tool.uv]
dev-dependencies = [
  "slowapi==0.1.9",
  "pytest==8.3.4",
  "pytest-asyncio==0.25.3",
  "httpx==0.27.2",
  "aiosqlite==0.21.0",
]
